+++ server/controllers/proximityAlert/crimeVersion.test.ts
@@ -140,15 +140,25 @@ describe('CrimeVersionController', () => {
                     latitude: 10.0,
                     longitude: 10.0,
                     sequenceLabel: 'A1',
-                    confidence: 10,
-                    capturedDateTime: '2025-01-01T00:00',
+                    precision: 10,
+                    capturedDateTime: '01/01/2025 00:00',
+                    deviceId: 1,
+                    direction: 0,
+                    name: 'name',
+                    nomisId: 'nomisId',
+                    speed: 0,
                   },
                   {
                     latitude: 10.0,
                     longitude: 10.0,
                     sequenceLabel: 'A2',
-                    confidence: 10,
-                    capturedDateTime: '2025-01-01T02:00',
+                    precision: 10,
+                    capturedDateTime: '01/01/2025 02:00',
+                    deviceId: 1,
+                    direction: 0,
+                    name: 'name',
+                    nomisId: 'nomisId',
+                    speed: 0,
                   },
                 ],
               },
@@ -156,45 +166,18 @@ describe('CrimeVersionController', () => {
           },
           versionColour: 'green',
           versionLabel: 'Latest version',
-        },
-        positions: [
-          {
-            latitude: 10,
-            longitude: 10,
-            positionType: 'crime',
-            crimeTypeId: 'AB',
+          crimePosition: {
             batchId: 'batch1',
             crimeDateTimeFrom: '01/01/2025 00:00',
-            crimeDateTimeTo: '01/01/2025 00:00',
+            crimeDateTimeTo: '01/01/2025 01:00',
             crimeReference: 'crime1',
+            latitude: 10,
+            longitude: 10,
             overlayBodyTemplateId: 'overlay-body-crime-location',
             overlayTitleTemplateId: 'overlay-title-crime-location',
+            precision: 100,
           },
-          {
-            precision: 10,
-            deviceId: 1,
-            geolocationMechanism: 'GPS',
-            latitude: 10,
-            longitude: 10,
-            personName: 'name',
-            personNomisId: 'nomisId',
-            positionType: 'wearer',
-            sequenceLabel: 'A1',
-            timestamp: '2025-01-01T00:00',
-          },
-          {
-            precision: 10,
-            deviceId: 1,
-            geolocationMechanism: 'GPS',
-            latitude: 10,
-            longitude: 10,
-            personName: 'name',
-            personNomisId: 'nomisId',
-            positionType: 'wearer',
-            sequenceLabel: 'A2',
-            timestamp: '2025-01-01T02:00',
-          },
-        ],
+        },
         exportProximityAlertForm: {
           url: `/proximity-alert/${crimeVersionId}/export-proximity-alert`,
           selectedDeviceIds: [],
@@ -259,21 +242,18 @@ describe('CrimeVersionController', () => {
           matching: { deviceWearers: [] },
           versionLabel: 'Latest version',
           versionColour: 'green',
-        },
-        positions: [
-          {
-            latitude: 10,
-            longitude: 10,
-            positionType: 'crime',
-            crimeTypeId: 'AB',
+          crimePosition: {
             batchId: 'batch1',
             crimeDateTimeFrom: '01/01/2025 00:00',
-            crimeDateTimeTo: '01/01/2025 00:00',
+            crimeDateTimeTo: '01/01/2025 01:00',
             crimeReference: 'crime1',
+            latitude: 10,
+            longitude: 10,
             overlayBodyTemplateId: 'overlay-body-crime-location',
             overlayTitleTemplateId: 'overlay-title-crime-location',
+            precision: 100,
           },
-        ],
+        },
         exportProximityAlertForm: {
           url: `/proximity-alert/${crimeVersionId}/export-proximity-alert`,
           selectedDeviceIds: [],
@@ -338,21 +318,18 @@ describe('CrimeVersionController', () => {
           matching: null,
           versionColour: 'green',
           versionLabel: 'Latest version',
-        },
-        positions: [
-          {
-            latitude: 10,
-            longitude: 10,
-            positionType: 'crime',
-            crimeTypeId: 'AB',
+          crimePosition: {
             batchId: 'batch1',
             crimeDateTimeFrom: '01/01/2025 00:00',
-            crimeDateTimeTo: '01/01/2025 00:00',
+            crimeDateTimeTo: '01/01/2025 01:00',
             crimeReference: 'crime1',
+            latitude: 10,
+            longitude: 10,
             overlayBodyTemplateId: 'overlay-body-crime-location',
             overlayTitleTemplateId: 'overlay-title-crime-location',
+            precision: 100,
           },
-        ],
+        },
         exportProximityAlertForm: {
           url: `/proximity-alert/${crimeVersionId}/export-proximity-alert`,
           selectedDeviceIds: [],
@@ -441,21 +418,18 @@ describe('CrimeVersionController', () => {
           matching: { deviceWearers: [] },
           versionLabel: 'Latest version',
           versionColour: 'green',
-        },
-        positions: [
-          {
-            latitude: 10,
-            longitude: 10,
-            positionType: 'crime',
-            crimeTypeId: 'AB',
+          crimePosition: {
             batchId: 'batch1',
             crimeDateTimeFrom: '01/01/2025 00:00',
-            crimeDateTimeTo: '01/01/2025 00:00',
+            crimeDateTimeTo: '01/01/2025 01:00',
             crimeReference: 'crime1',
+            latitude: 10,
+            longitude: 10,
             overlayBodyTemplateId: 'overlay-body-crime-location',
             overlayTitleTemplateId: 'overlay-title-crime-location',
+            precision: 100,
           },
-        ],
+        },
         exportProximityAlertForm: {
           url: `/proximity-alert/${crimeVersionId}/export-proximity-alert`,
           selectedDeviceIds: ['1'],
